requirements.txt
exceptiongroup==1.2.2
ibm_db==3.2.5
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
PyYAML==6.0.2
tomli==2.2.1

pytest-bdd~=8.1.0
pytest
pip~=22.0.4
Mako~=1.3.8
MarkupSafe~=3.0.2
gherkin-official~=29.0.0
parse-type~=0.6.4
parse~=1.20.2
six~=1.17.0
setuptools~=58.1.0
ibm-db~=3.2.3
boto3==1.36.8
botocore==1.36.8
s3transfer==0.11.2
mypy-boto3-s3~=1.36.9